+++ changelog.md
@@ -8,6 +8,7 @@
 - structurizr-client: Removes `StructurizrClient` (use `WorkspaceApiClient` instead).
 - structurizr-import: Adds support for importing decisions managed by Log4brains.
 - structurizr-import: Adds support for importing decisions in MADR format.
+- structurizr-import: Fixes https://github.com/structurizr/java/issues/251 (Importing docs fails on files without extension).
 - structurizr-dsl: Adds `!decisions` as a synonym for `!adrs`.
 - structurizr-dsl: Allows `!identifiers` to be used inside `model`.
 - structurizr-inspection: Initial version.